Feature/s1 news

See merge request emtiv/backend/emtiv-backend!297

+++ src/services/newsService.js
@@ -1,12 +1,29 @@
 const News = require('../models/News')
 const NewsView = require('../models/NewsView')
 const CompanyUserMapping = require('../models/CompanyUserMapping')
+const ModuleNameEnum = require('../models/enum/ModuleNameEnum')
 const { UnsupportedOperationError, NotFoundError } = require('../models/errors')
 const ServiceHelper = require('../helper/ServiceHelper')
+const industryService = require('./industryService')
+const settingService = require('./settingService')
+const gradeService = require('./gradeService')
+const companyService = require('./companyService')
+const fileService = require('./fileService')
 
 const UnsupportedOperationErrorEnum = {
     USER_NOT_IN_COMPANY: 'USER_NOT_IN_COMPANY',
-    NEWS_NOT_EXIST: 'NEWS_NOT_EXIST'
+    NEWS_NOT_EXIST: 'NEWS_NOT_EXIST',
+    INDUSTRY_NOT_FOUND: 'INDUSTRY_NOT_FOUND',
+    FORBIDDEN_ACTION: 'FORBIDDEN_ACTION',
+    FILE_NOT_EXIST: 'FILE_NOT_EXIST',
+    NEWS_ALREADY_PUBLISHED: 'NEWS_ALREADY_PUBLISHED',
+    INVALID_TYPE: 'INVALID_TYPE'
+}
+
+const NewsTypeEnum = {
+    PUBLISHED: 'PUBLISHED',
+    UNPUBLISHED: 'UNPUBLISHED',
+    DRAFT: 'DRAFT'
 }
 
 const newsService = {}
@@ -28,83 +45,118 @@ newsService.getNewsByIdAndCompanyId = async (newsId, user) => {
 
     return News.query()
     .findById(newsId)
+    .modify('baseAttributes')
     .where('ecompanyecompanyid', user.companyId)
-    .where('enewsispublished', true)
-    .then(news => {
-        if(!news) return false
-        return news
-    })
 
 }
 
 newsService.createNews = async (newsDTO, user) => {
 
-    const userInCompany = await newsService.userInCompany(user)
+    const userInCompany = await companyService.isUserExistInCompany(newsDTO.ecompanyecompanyid, user.sub)
 
     if(!userInCompany) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.USER_NOT_IN_COMPANY)
 
-    return News.transaction(trx => {
-        return News.query(trx)
-        .insertToTable(newsDTO, user.sub)
-    })
+    const isAllowed = await gradeService.getAllGradesByUserIdAndCompanyId(newsDTO.ecompanyecompanyid, user.sub)
+        .then(grades => grades.map(grade => grade.egradeid))
+        .then(gradeIds => settingService.isUserHaveFunctions(['C'], gradeIds, ModuleNameEnum.NEWS, newsDTO.ecompanyecompanyid))
+
+    if (!isAllowed) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.FORBIDDEN_ACTION)
+
+    const industry = await industryService.getIndustryById(newsDTO.eindustryeindustryid)
+
+    if (!industry) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.INDUSTRY_NOT_FOUND)
+
+    const file = await fileService.getFileByIdAndCreateBy(newsDTO.efileefileid, user.sub)
+
+    if (!file) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.FILE_NOT_EXIST)
+
+    return News.query().insertToTable(newsDTO, user.sub)
 
 }
 
 newsService.publishNews = async (isPublish, newsId, user) => {
 
-    const userInCompany = await newsService.userInCompany(user)
+    const news = await News.query()
+        .where('enewsid', newsId)
+        .where('ecompanyecompanyid', user.companyId)
+        .first()
+
+    if (!news) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.NEWS_NOT_EXIST)
+
+    const userInCompany = await companyService.isUserExistInCompany(news.ecompanyecompanyid, user.sub)
 
     if(!userInCompany) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.USER_NOT_IN_COMPANY)
 
-    return News.transaction(trx => {
-        return News.query(trx)
-        .where('enewsid', newsId)
-        .where('ecompanyecompanyid', user.companyId)
-        .updateByUserId({enewsispublished: isPublish}, user.sub)
-        .first()
+    if (news.enewsispublished) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.NEWS_ALREADY_PUBLISHED)
+
+    if (news.enewsispublished && news.enewsispublished !== isPublish) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.NEWS_ALREADY_PUBLISHED)
+
+    const isAllowed = await gradeService.getAllGradesByUserIdAndCompanyId(news.ecompanyecompanyid, user.sub)
+        .then(grades => grades.map(grade => grade.egradeid))
+        .then(gradeIds => settingService.isUserHaveFunctions(['P'], gradeIds, ModuleNameEnum.NEWS, news.ecompanyecompanyid))
+
+    if (!isAllowed) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.FORBIDDEN_ACTION)
+
+    return news.$query()
+        .updateByUserId({ enewsispublished: isPublish, enewsdate: Date.now() }, user.sub)
         .returning('*')
-        .then(news => {
-            if(!news) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.NEWS_NOT_EXIST)
-            return news
-        })
-    })
 
 }
 
 newsService.editNews = async (newsDTO, newsId, user) => {
 
-    const userInCompany = await newsService.userInCompany(user)
+    const news = await News.query()
+        .where('enewsid', newsId)
+        .where('ecompanyecompanyid', user.companyId)
+        .first()
+
+    if (!news) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.NEWS_NOT_EXIST)
+
+    const userInCompany = await companyService.isUserExistInCompany(news.ecompanyecompanyid, user.sub)
 
     if(!userInCompany) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.USER_NOT_IN_COMPANY)
 
-    return News.query()
-    .where('enewsid', newsId)
-    .where('ecompanyecompanyid', user.companyId)
+    if (news.enewsispublished) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.NEWS_ALREADY_PUBLISHED)
+
+    const isAllowed = await gradeService.getAllGradesByUserIdAndCompanyId(news.ecompanyecompanyid, user.sub)
+        .then(grades => grades.map(grade => grade.egradeid))
+        .then(gradeIds => settingService.isUserHaveFunctions(['U'], gradeIds, ModuleNameEnum.NEWS, news.ecompanyecompanyid))
+
+    if (!isAllowed) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.FORBIDDEN_ACTION)
+
+    const industry = await industryService.getIndustryById(newsDTO.eindustryeindustryid)
+
+    if (!industry) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.INDUSTRY_NOT_FOUND)
+
+    if (newsDTO.efileefileid !== news.efileefileid) {
+
+        const file = await fileService.getFileByIdAndCreateBy(newsDTO.efileefileid, user.sub)
+
+        if (!file) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.FILE_NOT_EXIST)
+
+        await fileService.deleteFileByIdAndCreateBy(news.efileefileid, user.sub)
+    }
+
+    return news.$query()
     .updateByUserId(newsDTO, user.sub)
-    .first()
     .returning('*')
-    .then(news => {
-        if(!news) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.NEWS_NOT_EXIST)
-        return news
-    })
 
 }
 
-newsService.getNews = async (page, size, user, type) => {
+newsService.getNews = async (page, size, user, type, companyId, isPublic) => {
 
-    if(type !== 'UNPUBLISHED' && type !== 'PUBLISHED')
-        throw new NotFoundError()
+    if (NewsTypeEnum.PUBLISHED !== type && NewsTypeEnum.UNPUBLISHED !== type && NewsTypeEnum.DRAFT !== type)
+        throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.INVALID_TYPE)
 
-    if(type === 'UNPUBLISHED')
-        return News.query()
-        .where('ecompanyecompanyid', user.companyId)
-        .where('enewsispublished', false)
-        .page(page, size)
-        .then(pageObj => ServiceHelper.toPageObj(page, size, pageObj))
-    
-    return News.query()
-    .where('ecompanyecompanyid', user.companyId)
-    .where('enewsispublished', true)
+    let query = News.query()
+        .modify('list')
+        .where('enewsispublished', NewsTypeEnum.UNPUBLISHED !== type)
+
+    if (companyId) query = query.where('ecompanyecompanyid', companyId)
+
+    if (isPublic !== undefined) query = query.where('enewsispublic', isPublic)
+
+    return query
     .page(page, size)
     .then(pageObj => ServiceHelper.toPageObj(page, size, pageObj)) 
 
@@ -116,23 +168,6 @@ newsService.getNewsDetail = async (newsId, user) => {
 
     if(!newsFromDB) throw new NotFoundError()
 
-    const isNewsViewed = await NewsView.query()
-    .where('enewsenewsid', newsId)
-    .where('eusereuserid', user.sub)
-    .first()
-
-    // if user haven't viewed this news
-    if(!isNewsViewed){
-
-            const newsViewDTO = {
-                enewsenewsid: newsId,
-                eusereuserid: user.sub,
-            }
-
-            await NewsView.query().insertToTable(newsViewDTO, user.sub)
-
-    } 
-
     return newsFromDB
 }
 
@@ -140,7 +175,7 @@ newsService.generateNewsLink = async (newsId, user) => {
 
     const newsFromDB = await newsService.getNewsByIdAndCompanyId(newsId, user)
 
-    if(!newsFromDB) throw new NotFoundError() 
+    if(!newsFromDB) throw new NotFoundError()
     
     return `https://org.sportiv.app/news/${newsId}`
     
@@ -158,16 +193,41 @@ newsService.getUserViewCount = async (newsId) => {
 
 newsService.deleteNews = async (newsId, user) => {
 
-    const userInCompany = await newsService.userInCompany(user)
+    const news = await News.query()
+        .where('enewsid', newsId)
+        .where('ecompanyecompanyid', user.companyId)
+        .first()
+
+    if (!news) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.NEWS_NOT_EXIST)
+
+    const userInCompany = await companyService.isUserExistInCompany(news.ecompanyecompanyid, user.sub)
 
     if(!userInCompany) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.USER_NOT_IN_COMPANY)
 
-    return News.query()
-    .findById(newsId)
-    .where('ecompanyecompanyid', user.companyId)
-    .delete()
-    .then(rowsAffected => rowsAffected === 1)
+    const isAllowed = await gradeService.getAllGradesByUserIdAndCompanyId(news.ecompanyecompanyid, user.sub)
+        .then(grades => grades.map(grade => grade.egradeid))
+        .then(gradeIds => settingService.isUserHaveFunctions(['D'], gradeIds, ModuleNameEnum.NEWS, news.ecompanyecompanyid))
 
+    if (!isAllowed) throw new UnsupportedOperationError(UnsupportedOperationErrorEnum.FORBIDDEN_ACTION)
+
+    return news.$query()
+    .delete()
+    .then(rowsAffected => fileService.deleteFileByIdAndCreateBy(news.efileefileid, user.sub)
+        .catch()
+        .then(() => rowsAffected === 1))
+
+}
+
+newsService.getNewsById = async (newsId) => {
+
+    console.log(newsId)
+
+    return News.query()
+        .findById(newsId)
+        .then(news => {
+            if (!news) throw new NotFoundError()
+            return news
+        })
 }
